requirements.txt
fastapi
uvicorn
langchain
chromadb
pypdf
sentence-transformers
torch
python-dotenv
requests
langchain-community